Cleaned up chat interface
