+++ FitGymTool.UI/src/app/models/interfaces/suggestion-item.interface.ts
@@ -1,0 +1,6 @@
+export interface SuggestionItem {
+  id?: string;
+  text: string;
+  icon?: string;
+  category?: string;
+}
